feat: rename hackday module -> engine
