+++ engine/__init__.py
@@ -1,15 +1,14 @@
 import os
 import tornado.web
 
-from hackday.view import ApiHandler, DefaultHandler
+from engine.view import ApiHandler, DefaultHandler
 
 
-class HackDay(tornado.web.Application):
+class Engine(tornado.web.Application):
     def __init__(self):
         self.handler = [
             (r"/api", ApiHandler),
-	]
-	
+	    ]
 
         self.settings = dict(
             title="fill_it",
@@ -23,9 +22,8 @@ class HackDay(tornado.web.Application):
             default_handler_class=DefaultHandler,
         )
 
-        super(HackDay, self).__init__(self.handler, **self.settings)
-
-app = HackDay()
+        super(Engine, self).__init__(self.handler, **self.settings)
+app = Engine()
 
 if __name__ == "__main__":
     pass
